defect-kb/requirements.txt
chromadb>=0.5.0
pydantic>=2.7.0
pyyaml>=6.0
sentence-transformers>=2.7.0

# Advanced: only needed for --input / --auto-retry / openai embedding
# openai>=1.30.0
# anthropic>=0.40.0
